# Openapi-Server-Stubs/OpenapiGenerator.Server/wwwroot/openapi-original.json
{
  "openapi" : "3.0.1",
  "info" : {
    "contact" : {
      "email" : "apiteam@swagger.io"
    },
    "description" : "This is a sample server Petstore server.  You can find out more about Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).  For this sample, you can use the api key `special-key` to test the authorization filters.",
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "termsOfService" : "http://swagger.io/terms/",
    "title" : "Swagger Petstore",
    "version" : "1.0.5"
  },
...
  "servers" : [ {
    "url" : "https://petstore.swagger.io/"
  }, {
    "url" : "http://petstore.swagger.io/"
  } ],
...
  "paths" : {
    "/pet/{petId}/uploadImage" : {
      "post" : {
        "operationId" : "uploadFile",
        "parameters" : [ {
          "description" : "ID of pet to update",
          "in" : "path",
          "name" : "petId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "properties" : {
                  "additionalMetadata" : {
                    "description" : "Additional data to pass to server",
                    "type" : "string"
                  },
                  "file" : {
                    "description" : "file to upload",
                    "format" : "binary",
                    "type" : "string"
                  }
                }
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponse"
                }
              }
            },
            "description" : "successful operation"
          }
        },
        "summary" : "uploads an image",
        "tags" : [ "pet" ]
      }
    },
    "/pet" : {
      "post" : {
        "operationId" : "addPet",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Pet"
              }
            }
          },
          "description" : "Pet object that needs to be added to the store",
          "required" : true
        },
        "responses" : {
          "405" : {
            "content" : { },
            "description" : "Invalid input"
          }
        },
        "summary" : "Add a new pet to the store",
        "tags" : [ "pet" ],
        "x-codegen-request-body-name" : "body"
      },
      "put" : {
        "operationId" : "updatePet",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Pet"
              }
            }
          },
          "description" : "Pet object that needs to be added to the store",
          "required" : true
        },
        "responses" : {
          "400" : {
            "content" : { },
            "description" : "Invalid ID supplied"
          },
          "404" : {
            "content" : { },
            "description" : "Pet not found"
          },
          "405" : {
            "content" : { },
            "description" : "Validation exception"
          }
        },
        "summary" : "Update an existing pet",
        "tags" : [ "pet" ],
        "x-codegen-request-body-name" : "body"
      }
    },
    "/pet/findByStatus" : {
      "get" : {
        "description" : "Multiple status values can be provided with comma separated strings",
        "operationId" : "findPetsByStatus",
        "parameters" : [ {
          "description" : "Status values that need to be considered for filter",
          "explode" : true,
          "in" : "query",
          "name" : "status",
          "required" : true,
          "schema" : {
            "items" : {
              "default" : "available",
              "enum" : [ "available", "pending", "sold" ],
              "type" : "string"
            },
            "type" : "array"
          },
          "style" : "form"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/Pet"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "successful operation"
          },
          "400" : {
            "content" : { },
            "description" : "Invalid status value"
          }
        },
        "summary" : "Finds Pets by status",
        "tags" : [ "pet" ]
      }
    },
    "/pet/findByTags" : {
      "get" : {
        "deprecated" : true,
        "description" : "Multiple tags can be provided with comma separated strings. Use tag1, tag2, tag3 for testing.",
        "operationId" : "findPetsByTags",
        "parameters" : [ {
          "description" : "Tags to filter by",
          "explode" : true,
          "in" : "query",
          "name" : "tags",
          "required" : true,
          "schema" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "style" : "form"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/Pet"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "successful operation"
          },
          "400" : {
            "content" : { },
            "description" : "Invalid tag value"
          }
        },
        "summary" : "Finds Pets by tags",
        "tags" : [ "pet" ]
      }
    },
    "/pet/{petId}" : {
      "delete" : {
        "operationId" : "deletePet",
        "parameters" : [ {
          "in" : "header",
          "name" : "api_key",
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Pet id to delete",
          "in" : "path",
          "name" : "petId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "400" : {
            "content" : { },
            "description" : "Invalid ID supplied"
          },
          "404" : {
            "content" : { },
            "description" : "Pet not found"
          }
        },
        "summary" : "Deletes a pet",
        "tags" : [ "pet" ]
      },
      "get" : {
        "description" : "Returns a single pet",
        "operationId" : "getPetById",
        "parameters" : [ {
          "description" : "ID of pet to return",
          "in" : "path",
          "name" : "petId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Pet"
                }
              }
            },
            "description" : "successful operation"
          },
          "400" : {
            "content" : { },
            "description" : "Invalid ID supplied"
          },
          "404" : {
            "content" : { },
            "description" : "Pet not found"
          }
        },
        "summary" : "Find pet by ID",
        "tags" : [ "pet" ]
      },
      "post" : {
        "operationId" : "updatePetWithForm",
        "parameters" : [ {
          "description" : "ID of pet that needs to be updated",
          "in" : "path",
          "name" : "petId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/x-www-form-urlencoded" : {
              "schema" : {
                "properties" : {
                  "name" : {
                    "description" : "Updated name of the pet",
                    "type" : "string"
                  },
                  "status" : {
                    "description" : "Updated status of the pet",
                    "type" : "string"
                  }
                }
              }
            }
          }
        },
        "responses" : {
          "405" : {
            "content" : { },
            "description" : "Invalid input"
          }
        },
        "summary" : "Updates a pet in the store with form data",
        "tags" : [ "pet" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "ApiResponse" : {
        "example" : {
          "code" : 0,
          "type" : "type",
          "message" : "message"
        },
        "properties" : {
          "code" : {
            "format" : "int32",
            "type" : "integer"
          },
          "type" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "Category" : {
        "example" : {
          "name" : "name",
          "id" : 6
        },
        "properties" : {
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "name" : {
            "type" : "string"
          }
        },
        "type" : "object",
        "xml" : {
          "name" : "Category"
        }
      },
      "Pet" : {
        "example" : {
          "photoUrls" : [ "photoUrls", "photoUrls" ],
          "name" : "doggie",
          "id" : 0,
          "category" : {
            "name" : "name",
            "id" : 6
          },
          "tags" : [ {
            "name" : "name",
            "id" : 1
          }, {
            "name" : "name",
            "id" : 1
          } ],
          "status" : "available"
        },
        "properties" : {
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "category" : {
            "$ref" : "#/components/schemas/Category"
          },
          "name" : {
            "example" : "doggie",
            "type" : "string"
          },
          "photoUrls" : {
            "items" : {
              "type" : "string",
              "xml" : {
                "name" : "photoUrl"
              }
            },
            "type" : "array",
            "xml" : {
              "wrapped" : true
            }
          },
          "tags" : {
            "items" : {
              "$ref" : "#/components/schemas/Tag"
            },
            "type" : "array",
            "xml" : {
              "wrapped" : true
            }
          },
          "status" : {
            "description" : "pet status in the store",
            "enum" : [ "available", "pending", "sold" ],
            "type" : "string"
          }
        },
        "required" : [ "name", "photoUrls" ],
        "type" : "object",
        "xml" : {
          "name" : "Pet"
        }
      },
      "Tag" : {
        "example" : {
          "name" : "name",
          "id" : 1
        },
        "properties" : {
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "name" : {
            "type" : "string"
          }
        },
        "type" : "object",
        "xml" : {
          "name" : "Tag"
        }
      },
...
    }
  },
...
}